Split bookmark sync responsibilities
