+++ lib/features/bookmarks/data/sync/bookmarks_sync_service.dart
@@ -7,10 +7,9 @@ import 'package:injectable/injectable.dart';
 import '../../../../core/future_extensions.dart';
 import '../../domain/services/bookmarks_sync_controller.dart';
 import '../datasources/bookmarks_remote_data_source.dart';
-import '../local/bookmark_entity.dart';
 import '../local/bookmarks_local_data_source.dart';
-import '../models/bookmark_dto.dart';
-import '../models/bookmark_request.dart';
+import 'bookmarks_pull_reconciler.dart';
+import 'bookmarks_push_queue.dart';
 
 /// Reconciles the local ObjectBox store with the remote API.
 ///
@@ -35,16 +34,16 @@ import '../models/bookmark_request.dart';
 /// owning reads.
 @LazySingleton(as: BookmarksSyncController)
 class BookmarksSyncService implements BookmarksSyncController {
-  BookmarksSyncService(this._local, this._remote, this._connectivity);
+  BookmarksSyncService(
+    BookmarksLocalDataSource local,
+    BookmarksRemoteDataSource remote,
+    this._connectivity,
+  ) : _pushQueue = BookmarksPushQueue(local, remote),
+      _pullReconciler = BookmarksPullReconciler(local, remote);
 
-  final BookmarksLocalDataSource _local;
-  final BookmarksRemoteDataSource _remote;
   final Connectivity _connectivity;
-
-  /// Bound on concurrent media uploads per row. Picked so the server isn't
-  /// swamped by a row with many attachments, while still finishing notably
-  /// faster than a serial drain.
-  static const int _maxUploadConcurrency = 3;
+  final BookmarksPushQueue _pushQueue;
+  final BookmarksPullReconciler _pullReconciler;
 
   final _status = StreamController<BookmarksSyncStatus>.broadcast();
   StreamSubscription<List<ConnectivityResult>>? _connectivitySub;
@@ -98,8 +97,8 @@ class BookmarksSyncService implements BookmarksSyncController {
   Future<void> _run() async {
     _emit(BookmarksSyncStatus.syncing);
     try {
-      final pushHadFailures = await _push();
-      await _pull();
+      final pushHadFailures = await _pushQueue.push();
+      await _pullReconciler.pull();
       // A failed row keeps its pending state and is retried next trigger; we
       // still surface `error` so the UI reflects that the sync was incomplete.
       _emit(
@@ -117,201 +116,5 @@ class BookmarksSyncService implements BookmarksSyncController {
   void _emit(BookmarksSyncStatus s) {
     _statusNow = s;
     _status.add(s);
-  }
-
-  /// Drains every pending row. Each row is isolated so a single permanently
-  /// rejected row (e.g. a server `400`/`500`) can't block the rest of the
-  /// queue — it stays pending and is retried next trigger. Returns `true` if
-  /// any row failed.
-  Future<bool> _push() async {
-    final pending = await _local.listPending();
-    var hadFailure = false;
-    for (final row in pending) {
-      try {
-        await _pushRow(row);
-      } on Object {
-        hadFailure = true;
-      }
-    }
-    return hadFailure;
-  }
-
-  Future<void> _pushRow(BookmarkEntity row) async {
-    switch (row.syncState) {
-      case SyncState.pendingCreate:
-        await _checkpointUploads(row);
-        try {
-          final dto = await _remote.create(
-            BookmarkRequest(
-              id: row.uuid,
-              title: row.title,
-              url: row.url,
-              description: row.description,
-              tags: row.tags,
-              imageUrls: row.imageUrls,
-              videoUrl: row.videoUrl,
-            ),
-          );
-          _markSynced(row, dto);
-          await _local.put(row);
-        } on DioException catch (e) {
-          // 409: a previous attempt already created this row server-side (its
-          // response was lost). The uuid exists remotely, so stop treating it
-          // as pending — `_pull` reconciles its fields.
-          if (e.response?.statusCode != 409) rethrow;
-          row.syncState = SyncState.synced;
-          await _local.put(row);
-        }
-      case SyncState.pendingUpdate:
-        await _checkpointUploads(row);
-        final dto = await _remote.update(
-          row.uuid,
-          BookmarkRequest(
-            title: row.title,
-            url: row.url,
-            description: row.description,
-            tags: row.tags,
-            imageUrls: row.imageUrls,
-            videoUrl: row.videoUrl,
-          ),
-        );
-        _markSynced(row, dto);
-        await _local.put(row);
-      case SyncState.pendingDelete:
-        try {
-          await _remote.delete(row.uuid);
-        } on DioException catch (e) {
-          // 404 is fine — server already lost it. Anything else, rethrow so the
-          // per-row guard leaves the row pending for the next sync.
-          if (e.response?.statusCode != 404) rethrow;
-        }
-        await _local.hardDelete(row.id);
-      case SyncState.synced:
-        break;
-    }
-  }
-
-  void _markSynced(BookmarkEntity row, BookmarkDto dto) {
-    row
-      ..syncState = SyncState.synced
-      ..serverUpdatedAt = dto.updatedAt
-      ..imageUrls = List.of(dto.imageUrls)
-      ..videoUrl = dto.videoUrl;
-  }
-
-  /// Uploads any local media files and persists the resulting remote URLs back
-  /// to the row *before* the create/update call. If that call later fails, the
-  /// next retry sees already-uploaded `http(s)` URLs and skips re-uploading,
-  /// avoiding duplicate/orphaned blobs on the server.
-  Future<void> _checkpointUploads(BookmarkEntity row) async {
-    final uploadedImages = await _uploadMediaFiles(row.imageUrls);
-    final uploadedVideo = await _uploadVideoFile(row.videoUrl);
-    if (_sameUrls(uploadedImages, row.imageUrls) &&
-        uploadedVideo == row.videoUrl) {
-      return;
-    }
-    row
-      ..imageUrls = uploadedImages
-      ..videoUrl = uploadedVideo;
-    await _local.put(row);
-  }
-
-  bool _sameUrls(List<String> a, List<String> b) {
-    if (a.length != b.length) return false;
-    for (var i = 0; i < a.length; i++) {
-      if (a[i] != b[i]) return false;
-    }
-    return true;
-  }
-
-  Future<void> _pull() async {
-    final serverList = await _remote.list();
-    final serverByUuid = {for (final dto in serverList) dto.id: dto};
-    final localByUuid = <String, BookmarkEntity>{
-      for (final row in await _local.listAll()) row.uuid: row,
-    };
-
-    // Inserts and updates from the server payload.
-    for (final dto in serverList) {
-      final local = localByUuid[dto.id];
-      if (local == null) {
-        await _local.put(
-          BookmarkEntity(
-            uuid: dto.id,
-            title: dto.title,
-            url: dto.url,
-            description: dto.description,
-            tags: List.of(dto.tags),
-            imageUrls: List.of(dto.imageUrls),
-            videoUrl: dto.videoUrl,
-            createdAt: dto.createdAt,
-            updatedAt: dto.updatedAt,
-            serverUpdatedAt: dto.updatedAt,
-            syncStateCode: SyncState.synced.code,
-          ),
-        );
-        continue;
-      }
-      // Don't clobber rows the user has touched since the last sync.
-      if (local.syncState.isPending) continue;
-      // Server has a strictly newer version → take it.
-      if (dto.updatedAt.isAfter(local.updatedAt)) {
-        local
-          ..title = dto.title
-          ..url = dto.url
-          ..description = dto.description
-          ..tags = List.of(dto.tags)
-          ..imageUrls = List.of(dto.imageUrls)
-          ..videoUrl = dto.videoUrl
-          ..updatedAt = dto.updatedAt
-          ..serverUpdatedAt = dto.updatedAt;
-        await _local.put(local);
-      }
-    }
-
-    // Server-side deletes: any local `synced` row missing from the payload
-    // was deleted on another device.
-    for (final local in localByUuid.values) {
-      if (local.syncState != SyncState.synced) continue;
-      if (serverByUuid.containsKey(local.uuid)) continue;
-      await _local.hardDelete(local.id);
-    }
-  }
-
-  Future<List<String>> _uploadMediaFiles(List<String> paths) async {
-    // Uploads run in bounded-parallel chunks to finish a multi-attachment
-    // row faster without flooding the server. Order is preserved.
-    final results = List<String?>.filled(paths.length, null);
-    for (var start = 0; start < paths.length; start += _maxUploadConcurrency) {
-      final end = (start + _maxUploadConcurrency).clamp(0, paths.length);
-      await Future.wait(<Future<void>>[
-        for (var i = start; i < end; i++)
-          _uploadSingleMedia(paths[i]).then((url) => results[i] = url),
-      ]);
-    }
-    return [for (final url in results) url!];
-  }
-
-  Future<String> _uploadSingleMedia(String path) async {
-    if (path.startsWith('http://') || path.startsWith('https://')) {
-      return path;
-    }
-    final file = await MultipartFile.fromFile(path);
-    final res = await _remote.upload(file);
-    final url = res['url'];
-    if (url == null || url.isEmpty) {
-      // Surface as DioException so the outer _run() catch leaves the row
-      // pending for retry instead of silently dropping the attachment.
-      throw DioException(
-        requestOptions: RequestOptions(path: '/api/upload'),
-        error: 'Upload response missing url for $path',
-      );
-    }
-    return url;
-  }
-
-  Future<String?> _uploadVideoFile(String? path) async {
-    if (path == null) return null;
-    return _uploadSingleMedia(path);
   }
 }
